utec_scheduler/requirements.txt
Django>=3.2,<4.0
psycopg2-binary>=2.9
gunicorn>=20.0
deap>=1.3
python-dotenv>=0.19
djangorestframework>=3.14
django-cors-headers>=3.14
django-filter>=21.1